+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added method `fatal(string|Error, object)` in Logger
 - Added option `level: string` in Logger constructor (default: `info`), it can be one of `debug`, `info`, `warn`, `error` or `fatal`
 - Added option `outputs: Array<function>` in Logger constructor (default: `[consoleOutput()]`)
+- Added option `defaultContext: null|object` in Logger constructor (default: `null`)
 - Upgraded dependencies
 
 ## v2.2.9